Rename Dashboard component to UserDashboard
