+++ app/client/src/routes/allRebates.tsx
@@ -11,7 +11,7 @@ import { Message } from "components/message";
 import { MarkdownContent } from "components/markdownContent";
 import { TextWithTooltip } from "components/tooltip";
 import { useContentData } from "components/app";
-import { useCsbData, useBapSamData } from "components/dashboard";
+import { useCsbData, useBapSamData } from "components/userDashboard";
 import { useUserState } from "contexts/user";
 import { useNotificationsDispatch } from "contexts/notifications";
